+++ nevergrad/optimization/oneshot.py
@@ -7,6 +7,7 @@ import numpy as np
 from scipy import stats
 from scipy.spatial import ConvexHull  # pylint: disable=no-name-in-module
 import nevergrad.common.typing as tp
+from nevergrad.parametrization import parameter as p
 from . import sequences
 from . import base
 from .base import IntOrParameter
@@ -84,8 +85,15 @@ def avg_of_k_best(archive: utils.Archive[utils.MultiValue], method: str = "dimfo
 
 
 class OneShotOptimizer(base.Optimizer):
-    # pylint: disable=abstract-method
     one_shot = True
+
+    def _internal_ask_candidate(self) -> p.Parameter:
+        out = self.parametrization.spawn_child()
+        with p.helpers.deterministic_sampling(out):
+            # the values will be sampled deterministically since there is no reason not to
+            # with one shot algorithms
+            out.set_standardized_data(self._internal_ask())
+        return out
 
 
 # Recentering or center-based counterparts of the original Nevergrad oneshot optimizers:
@@ -110,7 +118,7 @@ class _RandomSearch(OneShotOptimizer):
         middle_point: bool = False,
         stupid: bool = False,
         opposition_mode: tp.Optional[str] = None,
-        cauchy: bool = False,
+        sampler: str = "parametrization",
         scale: tp.Union[float, str] = 1.0,
         recommendation_rule: str = "pessimistic",
     ) -> None:
@@ -121,8 +129,8 @@ class _RandomSearch(OneShotOptimizer):
         self.opposition_mode = opposition_mode
         self.stupid = stupid
         self.recommendation_rule = recommendation_rule
-        self.cauchy = cauchy
         self.scale = scale
+        self.sampler = sampler
         self._opposable_data: tp.Optional[np.ndarray] = None
 
     def _internal_ask(self) -> tp.ArrayLike:
@@ -144,11 +152,15 @@ class _RandomSearch(OneShotOptimizer):
             scale = np.sqrt(np.log(self.budget) / self.dimension)
         if isinstance(scale, str) and scale == "random":
             scale = np.exp(self._rng.normal(0.0, 1.0) - 2.0) / np.sqrt(self.dimension)
-        point = (
-            self._rng.standard_cauchy(self.dimension)
-            if self.cauchy
-            else self._rng.normal(0, 1, self.dimension)
-        )
+        # sample the new point
+        if self.sampler == "gaussian":
+            point = self._rng.normal(0, 1, self.dimension)
+        elif self.sampler == "cauchy":
+            point = self._rng.standard_cauchy(self.dimension)
+        elif self.sampler == "parametrization":
+            point = self.parametrization.sample().get_standardized_data(reference=self.parametrization)
+        else:
+            raise ValueError("Unkwnown sampler {self.sampler}")
         self._opposable_data = scale * point
         return self._opposable_data  # type: ignore
 
@@ -178,7 +190,11 @@ class RandomSearchMaker(base.ConfiguredOptimizer):
         symmetrizes exploration wrt the center: (e.g. https://ieeexplore.ieee.org/document/4424748)
              - full symmetry if "opposite"
              - random * symmetric if "quasi"
-    cauchy: bool
+    sampler: str
+        - parametrization: uses the default sample() method of the parametrization, which samples uniformly
+          between bounds and a Gaussian otherwise
+        - gaussian: uses a Gaussian distribution
+        - cauchy: uses a Cauchy distribution
         use a Cauchy distribution instead of Gaussian distribution
     scale: float or "random"
         scalar for multiplying the suggested point values, or string:
@@ -199,10 +215,11 @@ class RandomSearchMaker(base.ConfiguredOptimizer):
         middle_point: bool = False,
         stupid: bool = False,
         opposition_mode: tp.Optional[str] = None,
-        cauchy: bool = False,
+        sampler: str = "parametrization",
         scale: tp.Union[float, str] = 1.0,
         recommendation_rule: str = "pessimistic",
     ) -> None:
+        assert sampler in ["gaussian", "cauchy", "parametrization"]
         super().__init__(_RandomSearch, locals())