+++ nevergrad/parametrization/helpers.py
@@ -3,6 +3,7 @@
 # This source code is licensed under the MIT license found in the
 # LICENSE file in the root directory of this source tree.
 
+import contextlib
 import itertools
 import random
 import typing as tp
@@ -10,6 +11,7 @@ import numpy as np
 from . import core
 from . import container
 from . import choice
+from . import _layering
 from . import data as pdata
 
 
@@ -71,6 +73,30 @@ def flatten_parameter(
             if not isinstance(y, (container.Container, core.Constant)) or isinstance(y, choice.BaseChoice)
         }
     return flat
+
+
+@contextlib.contextmanager
+def deterministic_sampling(parameter: core.Parameter) -> tp.Iterator[None]:
+    """Temporarily change the behavior of a Parameter to become deterministic
+
+    Parameters
+    ----------
+    parameter: Parameter
+        the parameter which must behave deterministically during the "with" context
+    """
+    all_params = flatten_parameter(parameter)
+    int_layers = list(
+        itertools.chain.from_iterable([_layering.Int.filter_from(x) for x in all_params.values()])
+    )
+    # record state and set deterministic to True
+    deterministic = [lay.deterministic for lay in int_layers]
+    for lay in int_layers:
+        lay.deterministic = True
+    yield
+    # sample and reset the previous behavior
+    parameter.value  # pylint: disable=pointless-statement
+    for lay, det in zip(int_layers, deterministic):
+        lay.deterministic = det
 
 
 # pylint: disable=too-many-locals